Austin contextmenu AutoTranslate Script (#91)
* Update README.md

* Create AutoTranslate with CogNative.bat

Push batch file which translates whatever you drop on top of it. Place file in CogNative root. You can put a shortcut to that file in %appdata%\Microsoft\Windows\SendTo to add it to the Send To context menu on Windows.

* Update README.md

Add note about autotranslate script

* Update file extension portion

I hardcoded the presumed-true statement that our output is always a .wav even if the inputs aren't, and updated the readme accordingly.

* Update README.md

Make Andrew's suggested changes

Co-authored-by: austin.currington <austin.currington@tamu.edu>
Co-authored-by: Andrew Fennell <andrew.fennell1999@me.com>